% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClopPear.r
\name{cloppearSnSp}
\alias{cloppearSnSp}
\title{Binomial confidence interval, Clopper-Pearson method.}
\usage{
cloppearSnSp(dat, alpha = 0.05, est.Sn = TRUE)
}
\arguments{
\item{dat}{\code{data.frame}  A data frame with a column for the experimental test results, a column for the infallible reference test results,  and a column for the corresponding count.  The column
name for the experimental test results must contatin 'exp' and the column name for the infallible reference test results must include 'ref'.  The counts should be the last column.}

\item{alpha}{Complement of confidence level.}

\item{est.Sn}{logical (TRUE/FALSE) Indicating if the sensitivity and its confidence interval should be supplied (TRUE) or if the specificity and its confidence interval should be supplied (FALSE).}
}
\value{
An object of type \code{cp} that extends \code{list}. \cr \cr
\describe{
\item{\strong{calcVal}} {Named vector of point estimates and estimated simulated intervals. See below.}
\item{\strong{data}} {Test and Total values of the data. See below.}
\item{\strong{alpha}} {Complement of the confidence interval as provided above.}
}
}
\description{
Evaluate binomial confidence interval using Clopper-Pearson method. A function written by CVB Statistics to estimate the sensitivity and specificity
of an experimental diagnostic test kit in accordance with \href{https://www.aphis.usda.gov/aphis/ourfocus/animalhealth/veterinary-biologics/biologics-regulations-and-guidance/ct_vb_statwi}{CVB STATWI0002}.
}
\section{If \code{est.Sn == TRUE}}{


\code{calcVal} is a list with the following elements \cr
\itemize{
\item{\strong{Sn}} Sensitivity estimate.
\item{\strong{Sn.LL}} Lower confidence limit for sensitivity.
\item{\strong{Sn.UL}} Upper confidence limit for sensitivity.
}
\code{data} is a list with the following elements \cr
\itemize{
\item{\strong{Test.Positive}} Number of experimental test positives.
\item{\strong{Total.Positive}} Total number of positive samples.
}
}

\section{If \code{est.Sn == FALSE}}{


\code{calcVal} is a list with the following elements \cr
\itemize{
\item{\strong{Sp}} Specificity estimate.
\item{\strong{Sp.LL}} Lower confidence limit for specificity.
\item{\strong{Sp.UL}} Upper confidence limit for specificity.
}

\code{data} is a list with the following elements \cr
\itemize{
\item{\strong{Test.Negative}} Number of experimental test negatives.
\item{\strong{Total.Negative}} Total number of negative samples.
}
}
\examples{
CP.Sn <- cloppearSnSp(dat = dat_infal, est.Sn = TRUE)
CP.Sn
# Sn = P(T+|D+): 0.987013 (95\% CI: 0.953876, 0.998423)
CP.Sp <- cloppearSnSp(dat = dat_infal, est.Sn = FALSE)
CP.Sp
# Sp = P(T-|D-): 0.970297 (95\% CI: 0.915643, 0.915643)
}
\author{
CVB Statistics \email{CVB.Data.Help@aphis.usda.gov}
}
\references{
Clopper CJ, Pearson ES, 1934. The use of confidence or fiducial limits illustrated in the case of the binomial. \emph{Biometrika} 26:404-413.
}

