% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_infal}
\alias{dat_infal}
\title{Counts data used in vignette "DiagTestKit GettingStarted" section 4}
\format{
A data frame with 4 rows and 3 variables:
\describe{
\item{Experimental}{Result of the experimental test
(positive or negative).}
\item{Reference}{Result of the reference test
(positive or negative).}
\item{Count}{Number of samples observed with the unique
testing combination.}
}
}
\usage{
dat_infal
}
\description{
Counts of the positive and negative results for experimental and reference
tests. A 2-state experimental test when an infallible reference test has been
used to determine the true disease status of each sample.
}
\keyword{datasets}
