% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data7}
\alias{data7}
\title{Data set for Example 7 in vignette "DiagTestKit Examples"}
\format{A data frame with 39 rows and five variables:
\describe{
  \item{Population}{population identifier (A, B).}
  \item{exp_result}{result of experimental test (positive, negative).}
  \item{ref1_result}{result of first reference test (positive, negative, suspect).}
  \item{ref2_result}{result of second reference test (positive, negative, suspect).}
  \item{count}{number of samples with the unique testing combination for the
  specific population.}
}}
\usage{
data7
}
\description{
Samples randomly selected from 2 populations tested with two 3–state
 reference test and a 2–state experimental test.
}
\keyword{datasets}
