% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiagTestKit.r
\docType{package}
\name{DiagTestKit-package}
\alias{DiagTestKit}
\alias{DiagTestKit-package}
\title{DiagTestKit: Functions used in evaluating sensitivity and specificity at CVB Statistics}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A package written by CVB Statistics to estimate the sensitivity and specificity of an experimental diagnostic test kit in accordance with CVB STATWI0002 supporting the 2018 revision to VSM 800.73. See https://www.aphis.usda.gov/aphis/ourfocus/animalhealth/veterinary-biologics/biologics-regulations-and-guidance/ct_vb_statwi for more information.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ABS-dev/DiagTestKit/}
  \item \url{https://abs-dev.github.io/DiagTestKit/}
  \item Report bugs at \url{https://github.com/ABS-dev/DiagTestKit/issues/}
}

}
\author{
\strong{Maintainer}: CVB Statistics \email{CVB.Data.Help@usda.gov} [former owner]

Authors:
\itemize{
  \item David Siev (until 2019)
  \item Marie Vendettuoli (until 2019)
  \item Thomas Kent (starting 2020)
}

}
\keyword{internal}
