% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_parm.r
\name{.beta_parm}
\alias{.beta_parm}
\title{Convert Beta Parameterizations}
\usage{
.beta_parm(b, to = "alpha.beta")
}
\arguments{
\item{b}{\code{vector} A named vector specifying non-NULL values for 2
parameters. e.g. c(alpha=NA, beta=NA, mu=.6, theta=NA, phi=1.6, sigma2=NA)
or just c(mu=.6, phi=1.6)}

\item{to}{Specification of desired parameters, options are one of
"alpha.beta", "mu.phi", "mu.theta" or "mu.sigma2".}
}
\value{
\code{vector} A named vector with values for the parameters specified
  in the "to" argument of the input.
}
\description{
Convert between the paramaterizations of a beta distribution.
}
\author{
\link{DiagTestKit-package}
}
