% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_counts.r
\name{.cell_counts}
\alias{.cell_counts}
\title{Obtain cell counts}
\usage{
.cell_counts(
  SnR,
  SpR,
  Prev,
  SnE,
  SpE,
  sus.perc,
  N_mat,
  nstates,
  suspect2staterows,
  X,
  Xpos,
  Xsus,
  Xneg,
  ncells,
  ntests
)
}
\arguments{
\item{SnR}{\code{data.frame}  Each column corresponds to one reference test. Row 1
contains the sensitivity for the reference test(s). Row 2 contains the
probability of a suspect result as a fraction of the non-correct test
result. This is a value between 0 and 1 (inclusive). Namely, P(T? | D+) =
\eqn{\psi} = \eqn{\delta} * (1 - \eqn{\pi}) where \eqn{\delta} is the
second row for a given column (reference test).  \eqn{\delta =
  \frac{\psi}{(1 - \pi)}}{\delta = \psi / (1 - \pi)}.  Use a zero for a
2-state test (i.e. no suspect region).}

\item{SpR}{\code{data.frame} Each column corresponds to one reference test. Row 1
contains the specificity for each reference test. Row 2 contains the
probability of a suspect result as a fraction of the non-correct test
result.  This is a value between 0 and 1 (inclusive). Namely, P(T? | D-) =
\eqn{\phi} = \eqn{\gamma} * (1 - \eqn{\theta}) where \eqn{\gamma} is the
second row for a given column (reference test). \eqn{\gamma =
  \frac{\phi}{(1 - \theta)}}{\gamma = \phi / (1 - \theta)}.  Use a zero for a
2-state test (i.e. no suspect region).}

\item{Prev}{\code{vector}  A named vector containing the prevalence for each
population sampled.}

\item{SnE}{Sensitivity of the experimental test kit.}

\item{SpE}{Specificity of the experimental test kit.}

\item{sus.perc}{\code{vector} A vector containing 2 elements, c(\eqn{\delta},
\eqn{\gamma}) for the experimental test kit.  A vector of zeros for a
2-state experimental kit. \eqn{\delta} and \eqn{\gamma} are values between
0 and 1 (inclusive) corresponding to the proportion of the remaining
probability (i.e. 1 - \eqn{\pi} or 1 - \eqn{\theta}) that is suspect
(\eqn{\psi} or \eqn{\phi}).  \eqn{\delta = \frac{\psi}{(1-\pi)}} and
\eqn{\gamma = \frac{\phi}{(1-\theta)}}.}

\item{N_mat}{\code{matrix} Needs to be filled out}

\item{nstates}{\code{vector} A vector with length one greater than the number of
reference tests.  The first element is the number of states of the
experimental test and the remaining entries are the number of states of
each reference test (using the same ordering as SnR and SpR).}

\item{suspect2staterows}{Needs to be filled out.}

\item{X}{Needs to be filled out.}

\item{Xpos}{Needs to be filled out.}

\item{Xsus}{Needs to be filled out.}

\item{Xneg}{Needs to be filled out.}

\item{ncells}{Needs to be filled out.}

\item{ntests}{Needs to be filled out.}
}
\value{
\code{vector} A vector of expected counts corresponding to the properties
of the reference and experimental tests.  The expected counts are obtained
based on a conditional independence assumption of all test methods.
}
\description{
This function creates expected cell counts (and probabilities)
for a specific test pattern based on the diagnostic characteristics of the
reference test(s) and experimental test.
}
\author{
\link{DiagTestKit-package}
}
