% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_triangle_dist.r
\name{.create_triangle_dist}
\alias{.create_triangle_dist}
\title{Create Triangular Distribution}
\usage{
.create_triangle_dist(
  m,
  w,
  h,
  threestate = FALSE,
  suspect = 2/3,
  step_size = 0.005,
  p_proportion = TRUE
)
}
\arguments{
\item{m}{This is a point estimate for the parameter in which you are
obtaining the distribution, e.g. sensitivity, specificity, or prevalence.}

\item{w}{\code{vector}  A vector that provides the half widths of the 3
regions, (w1 closest, w3 farthest).}

\item{h}{\code{vector}  A vector of "y" (pseudo-value until scaled to be a
probability) corresponding to the height of the shoulder and the height of
the plateau.}

\item{threestate}{logical (TRUE/FALSE) Indicates whether or not there is a
"suspect" region (i.e. positive/suspect/negative).}

\item{suspect}{A fraction that indicates what percentage of the remaining
probability would be assigned to the suspect region.  For instance, if the
function gives sensitivity and then the probability of "suspect" is (1 -
sensitivity) * suspect.}

\item{step_size}{distance between the 'x' in the discrete distribution,
resolution of possible observations of the created distribution.}

\item{p_proportion}{whether to express 'p' as a proportion of its sum.}
}
\value{
\code{data.frame} of 'x', 'y', and 'p'.
}
\description{
Creates a discrete step/triangular distribution that can be used
  to sample values for sensitivity or specificity of a reference test or
  prevalence of a population.
}
\author{
\link{DiagTestKit-package}
}
