% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_hpd.r
\name{.emp_hpd}
\alias{.emp_hpd}
\title{Calculate the empirical hpd.}
\usage{
.emp_hpd(X, alpha)
}
\arguments{
\item{X}{vector of values}

\item{alpha}{1 - confidence}
}
\value{
highest posterior density (1-alpha) interval
}
\description{
Empirical highest posterior density by shortest length interval.
}
\note{
Uses type 7 \code{\link{quantile}}. Also used in package \code{MF}
}
\author{
\link{DiagTestKit-package}
}
