% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_simulated_values.r
\name{.get_simulated_values}
\alias{.get_simulated_values}
\title{Get Simulated Values}
\usage{
.get_simulated_values(means, distn, spread, nsim, step_size, prevalence)
}
\arguments{
\item{means}{\code{vector} A named vector containing point estimates for the
prevalence of each population (when prevalence = TRUE, see below) or a data
frame where each column corresponds to a reference test and the rows are
sensitivity (\eqn{\pi}) and \eqn{\psi} (or specificity (\eqn{\theta})  and
\eqn{\phi}).}

\item{distn}{\code{vector}  A vector of same length as \code{means}. Values
may be one of \code{NULL}, 'beta', or 'triangular'. \code{NULL} will be
treated as 'beta'}

\item{spread}{\code{vector} A vector of same length as \code{means}. Values
may be one of \code{NULL}, 'wide', 'medium', or 'narrow'. \code{NULL} will
be treated as 'wide'.}

\item{nsim}{The number of simulations to draw from the sensitivity and
specificity distribution(s) for each reference test or the prevalence
distribution from each population.}

\item{step_size}{Provides the level of resolution in values simulated from a
triangular distribution.}

\item{prevalence}{logical (TRUE/FALSE)  TRUE indicates that the function is
simulating values of prevalence.  This will determine the structure of the
output.}
}
\value{
final_mat A matrix of simulated values.  If prevalence is TRUE,
  final_mat will have the number of columns corresponding to the number of
  populations sampled else if prevalence is FALSE, final_mat will have number
  of columns twice the number of reference tests.  The columns are
  sensitivity (or specificity) of the first reference test, the probability
  of a suspect result as a fraction of the non-correct test result (i.e.
  either \eqn{\delta} or \eqn{\gamma}) for the first reference and continues
  in the same pattern for all reference tests.
}
\description{
Simulate values for use in optimization.  This function is used
  to obtain draws from a distribution for the sensitivity and specificity for
  each reference test and for the prevalence of each population tested.
}
\author{
\link{DiagTestKit-package}
}
