% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_value.r
\name{.get_values}
\alias{.get_values}
\title{Optimization of Sensitivity and Specificity}
\usage{
.get_values(
  dat,
  SnR.vec,
  SpR.vec,
  prev.vec,
  N.vec,
  nstates,
  tolerance,
  rep.iter,
  iter.n,
  parm = NULL
)
}
\arguments{
\item{dat}{\code{vector} A vector of counts ordered in a manner consistent
with output from the .cell_counts function.}

\item{SnR.vec}{\code{data.frame}  Each column corresponds to one reference
test.  Row 1 contains the sensitivity for the reference test(s). Row 2
contains the probability of a suspect result as a fraction of the
non-correct test result. This is a value between 0 and 1 (inclusive).
Namely, P(T? | D+) = \eqn{\psi} = \eqn{\delta} * (1 - \eqn{\pi}) where
\eqn{\delta} is the second row for a given column (reference test).
\eqn{\delta = \frac{\psi}{(1 - \pi)}}{\delta = \psi/(1 - \pi)}.  Use a zero
for a 2-state test (i.e. no suspect region).}

\item{SpR.vec}{\code{data.frame} Each column corresponds to one reference
test.  Row 1 contains the specificity for each reference test. Row 2
contains the probability of a suspect result as a fraction of the
non-correct test result.  This is a value between 0 and 1 (inclusive).
Namely, P(T? | D-) = \eqn{\phi} = \eqn{\gamma} * (1 - \eqn{\theta}) where
\eqn{\gamma} is the second row for a given column (reference test).
\eqn{\gamma = \frac{\phi}{(1 - \theta)}}{\gamma = \phi/(1 - \theta)}.  Use
a zero for a 2-state test (i.e. no suspect region).}

\item{prev.vec}{\code{vector}  A named vector containing the prevalence for
each population sampled.}

\item{N.vec}{\code{vector}  A named vector containing the sample size for
each population sampled.}

\item{nstates}{\code{vector} A vector with length one more than the number of
reference tests.  The first element is the number of states of the
experimental test and the remaining entries are the number of states of
each reference test (using the same ordering as SnR.vec and SpR.vec).}

\item{tolerance}{Setting a limit on the pgtol used in the optim function with
the 'L-BFGS-B' method. See also \code{\link{optim}}.}

\item{rep.iter}{logical (TRUE/FALSE)  Indicates if updates should be printed
regarding the number of iterations completed.}

\item{iter.n}{integer indicating the frequency of updates for the number of
iterations completed.}

\item{parm}{\code{vector}  A vector of starting values to be used for the
optimization that is passed to \code{.minimize_cell}.  For a 2-state
experimental test, this is a vector of length 2 with entries (\eqn{\pi},
\eqn{\theta}). For a 3-state experimental test, this is a vector of length
4 with entries (\eqn{\pi}, \eqn{\delta}, \eqn{\theta}, \eqn{\gamma}). See
also \code{\link{estimateSnSp}}.}
}
\value{
A list: \cr \cr The following will be returned for both 2 and 3-state
experimental tests -- \cr
\itemize{
\item{sens.final}  \code{vector} The optimized values for the sensitivity of
the experimental test kit.
\item{spec.final}  \code{vector} The optimized values for the specificity of
the experimental test kit.
\item{converge}  \code{vector} Each entry is an integer code detailing the
convergence of the optimization for each iteration.  0 indicates successful
completion. See also \code{\link{optim}}.
\item{message}  \code{vector}  Each entry includes a character string giving
any additional information returned by the optimizer or NULL.  See also
\code{\link{optim}}.
}

If three states -- \cr
\itemize{
\item{\eqn{\delta}}  \code{vector} The optimized values for the probability
of a suspect result as a fraction of the non-correct test result for
diseased samples.
\item{\eqn{\gamma}} \code{vector} The optimized value for the probability
of a suspect result as a fraction of the non-correct test result for
non-diseased samples.
}
}
\description{
Determine final optimized values for the sensitivity and
specificity of an experimental test kit (and probability of suspect given
disease positive and given disease negative for a 3-state kit).
}
\author{
\link{DiagTestKit-package}
}
