% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimize_cell.r
\name{.minimize_cell}
\alias{.minimize_cell}
\title{minimize cell}
\usage{
.minimize_cell(
  parm,
  SnR,
  SpR,
  Prev,
  xdat,
  N_mat,
  nstates,
  suspect2staterows,
  X,
  Xpos,
  Xsus,
  Xneg,
  ncells,
  ntests
)
}
\arguments{
\item{parm}{\code{vector}   A vector of starting values to be used for the
optimization that is passed to \code{.minimize_cell}.  For a 2-state
experimental test, this is a vector of length 2 with entries (\eqn{\pi},
\eqn{\theta}) For a 3-state experimental test, this is a vector of length 4
with entries (\eqn{\pi}, \eqn{\delta}, \eqn{\theta}, \eqn{\gamma}). See
also \code{\link{estimateSnSp}}.}

\item{SnR}{\code{data.frame}  Each column corresponds to one reference test.
Row 1 contains the sensitivity for the reference test(s). Row 2 contains
the probability of a suspect result as a fraction of the non-correct test
result. This is a value between 0 and 1 (inclusive). Namely, P(T? | D+) =
\eqn{\psi} = \eqn{\delta} * (1 - \eqn{\pi}) where \eqn{\delta} is the
second row for a given column (reference test).  \eqn{\delta =
\frac{\psi}{(1 - \pi)}}{\delta = \psi/(1 - \pi)}.  Use a zero for a 2-state
test (i.e. no suspect region).}

\item{SpR}{\code{data.frame} Each column corresponds to one reference test.
Row 1 contains the specificity for each reference test. Row 2 contains the
probability of a suspect result as a fraction of the non-correct test
result.  This is a value between 0 and 1 (inclusive). Namely, P(T? | D-) =
\eqn{\phi} = \eqn{\gamma} * (1 - \eqn{\theta}) where \eqn{\gamma} is the
second row for a given column (reference test). \eqn{\gamma =
\frac{\phi}{(1 - \theta)}}{\gamma = \phi/(1 - \theta)}.  Use a zero for a
2-state test (i.e. no suspect region).}

\item{Prev}{\code{vector}  A named vector containing the prevalence for each
population sampled.}

\item{xdat}{\code{vector}  A vector of the observed cell counts.}

\item{N_mat}{\code{matrix}  Needs to be filled out}

\item{nstates}{\code{vector} A vector with length one more than the number of
reference tests.  The first element is the number of states of the
experimental test and the remaining entries are the number of states of
each reference test (using the same ordering as SnR and SpR).}

\item{suspect2staterows}{Needs to be filled out.}

\item{X}{Needs to be filled out.}

\item{Xpos}{Needs to be filled out.}

\item{Xsus}{Needs to be filled out.}

\item{Xneg}{Needs to be filled out.}

\item{ncells}{Needs to be filled out.}

\item{ntests}{Needs to be filled out.}
}
\value{
The sum of the squared deviations between the expected and observed
  cell counts.
}
\description{
A function used for optimizing the values of sensitivity and
  specificity (and \eqn{\delta} and \eqn{\gamma} for a 3-state kit). The
  objective function minimizes the sum of the squared deviations (expected -
  observed cell counts).
}
\author{
\link{DiagTestKit-package}
}
