% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{estimateSnSpControl}
\alias{estimateSnSpControl}
\title{control values for estimateSnSp}
\usage{
estimateSnSpControl(seed = NULL, Sn.distn = NULL, Sn.spread = NULL,
  Sp.distn = NULL, Sp.spread = NULL, prev.distn = NULL,
  prev.spread = NULL, tolerance = 0.001, alpha = 0.05,
  step.size = 1e-06, parm = NULL, rep.iter = TRUE, iter.n = 50)
}
\arguments{
\item{seed}{The seed used in the random generation of the distributions of sensitivity and specificity for all reference tests and
prevalence of each population.  See also \code{\link{set.seed}}.}

\item{Sn.distn}{\code{vector}  A named vector with length equal to the number of reference tests.  Determines which disibution should be used for sampling sensitivity of each reference test.  Inputs are 'beta' or 'triangular'.
Defaults to 'beta' for each reference test.}

\item{Sn.spread}{\code{vector}  A named vector with length equal to the number of reference tests. Describes the width of the distribution for the sensitivity of each reference test.  Inputs are 'wide', 'medium',
or 'narrow'.  Defaults to 'wide' for each reference test.}

\item{Sp.distn}{\code{vector}  A named vector with length equal to the number of reference tests.  Determines which disibution should be used for sampling specificity of each reference test.  Inputs are 'beta' or 'triangular'.
Defaults to 'beta' for each reference test.}

\item{Sp.spread}{\code{vector}  A named vector with length equal to the number of reference tests. Describes the width of the distribution for the specificity of each reference test.  Inputs are 'wide', 'medium',
or 'narrow'.  Defaults to 'wide' for each reference test.}

\item{prev.distn}{\code{vector}  A named vector with length equal to the number of populations. Determines which disibution should be used for sampling the prevalence of each population.  Inputs are 'beta' or 'triangular'.
Defaults to 'beta'.}

\item{prev.spread}{\code{vector}  A named vector with length equal to the number of populations. Describes the width of the distribution for the prevalence of each population.  Inputs are 'wide', 'medium', or 'narrow'.
Defaults to 'wide' for each population.}

\item{tolerance}{Setting a limit on the pgtol used in the \code{\link{optim}} function with the 'L-BFGS-B' method. See also \code{\link{optim}}.  Defaults to 1E-03.}

\item{alpha}{Significance levels.  Defaults to 0.05.}

\item{step.size}{Provides the level of resolution in values simulated from a triangular distribution.  Defaults to 1E-06.}

\item{parm}{\code{vector}  Starting values for the optimization of the parameters of the experimental test.  If the experimental test has 2 states, this vector is of length two with elements corresponding to
sensitivity and specificity, respectively.  If the experimental test has 3 states, this vector is of length 4 with elements corresponding to sensitivity (\eqn{\pi}), the proportion of 1-Sn corresponding to the suspect region for disease
positive samples (\eqn{\delta}), specificity (\eqn{\theta}), and the proportion of 1-Sp corresponding to the suspect region for disease negative samples (\eqn{\gamma}).  All values are between 0 and 1, inclusive.}

\item{rep.iter}{logical (TRUE/FALSE) Indicates if updates should be printed regarding the number of iterations completed.  Defaluts to TRUE.}

\item{iter.n}{integer indicating the frequency of updates for the number of iterations completed.  Defaluts to 50.}
}
\value{
A list with the following elements (as defined above): \code{seed, Sn.disn, Sn.spread,
Sp.distn,} \code{ Sp.spread, prev.distn, prev.spread, tolerance, step.size, parm}
}
\description{
The values supplied in the function--call replace the defaults and a list with all
possible arguments is returned. The returned list is used as the \code{control} argument to
the function \code{estimateSnSp}.
}
\examples{
estimateSnSpControl()
estimateSnSpControl(seed = 64725)
}
\author{
CVB Statistics \email{CVB.Data.Help@aphis.usda.gov}
}

