% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data5}
\alias{data5}
\title{Data set for Example 5 in vignette "DiagTestKit Examples"}
\format{
A data frame with 11 rows and four variables:
\describe{
\item{Population}{population identifier (A, B).}
\item{exp_result}{result of experimental test
(positive, negative, suspect).}
\item{ref_result}{result of reference test (positive, negative).}
\item{count}{number of samples with the unique testing combination for the
specific population.}
}
}
\usage{
data5
}
\description{
Samples randomly selected from 2 populations tested by one 2–state
reference test and a 3–state experimental test.
}
\keyword{datasets}
