% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiagTestKit-package.r
\name{DiagTestKit-package}
\alias{DiagTestKit-package}
\title{DiagTestKit Package.}
\description{
A package written by CVB Statistics to estimate the sensitivity and 
specificity of an experimental diagnostic test kit in accordance with CVB STATWI0002
supporting the 2018 revision to VSM 800.73.
}
\details{
Functions used in evaluating sensitivity and specificity at CVB Statistics

\tabular{ll}{
Package: \tab DiagTestKit-package\cr
Type: \tab Package\cr
Version: \tab 0.5.3\cr
Date: \tab 26 July 2018\cr
License: \tab MIT\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
}
}
\section{Resources}{


\itemize{
\item GUIDANCE: \url{https://www.aphis.usda.gov/aphis/ourfocus/animalhealth/veterinary-biologics/biologics-regulations-and-guidance/ct_vb_statwi}
\item QUICK START: \url{https://github.com/ABS-dev/DiagTestKit/blob/master/README.md}
\item BUG REPORTS: \url{https://github.com/ABS-dev/DiagTestKit/issues}
}
}

\author{
CVB Statistics \email{CVB.Data.Help@aphis.usda.gov}
}
